Rename baseClients files
